+++ WHATS_NEW.md
@@ -8,11 +8,15 @@ New in 5.1
 Student Fields:
 - Search Medical Immunization or Physical, sponsored by Asian Hope
 
+Graes:
+- Rename "Edit Student Grades" program to "Historical Grades".
+
 Student Billing:
 - Payments: Fees dropdown to reconcile Payment.
 
 Accounting:
 - Staff Payments: Salaries dropdown to reconcile Payment.
+
 
 New in 5.0
 ----------