+++ WHATS_NEW.md
@@ -17,6 +17,8 @@ Student Billing:
 Accounting:
 - Staff Payments: Salaries dropdown to reconcile Payment.
 
+[Siblings](https://www.rosariosis.org/siblings-plugin/) plugin, sponsored by Asian Hope
+
 [REST API](https://gitlab.com/francoisjacquet/REST_API/) plugin